Adds text-to-speech functionality
Adds the ability to speak the input text in the chat list.

This change introduces a "Volume2" icon next to the input text in the chat list. Clicking this icon will trigger the text-to-speech functionality, reading the input text aloud.

The `speak` function now passes the text to the `detect_language` invoke, allowing the function to accurately determine language. Also, `speak` now includes a console log to aid with debug.